chatbot/requirements.txt
python-dotenv
requests
websockets
isort>=5.12.0
black>=23.3.0
flake8>=6.0.0